backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
pydantic==2.13.3
pdfplumber==0.11.6
python-multipart==0.0.20
python-docx
openpyxl
